requirements.txt
smart-open==7.0.4
boto3==1.35.10
mysql-connector-python==9.0.0